+++ tests/Fixtures/Arch/ToBeInvokable/IsInvokable/InvokableClassViaParent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\Arch\ToBeInvokable\IsInvokable;
+
+class InvokableClassViaParent extends ParentInvokableClass
+{
+    //
+}