+++ tests/Fixtures/Arch/ToBeInvokable/IsInvokable/InvokableTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\Arch\ToBeInvokable\IsInvokable;
+
+trait InvokableTrait
+{
+    public function __invoke(): void
+    {
+        //
+    }
+}